requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
markdown>=3.5.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
cryptography>=42.0.0
alembic>=1.13.0
boto3>=1.34.0
pgvector>=0.2.5

# LangChain Agentic RAG Dependencies
langchain>=0.2.0,<0.4.0
langchain-core>=0.2.0,<0.4.0
langchain-openai>=0.1.0,<0.3.0
langchain-community>=0.2.0,<0.4.0
langchain-text-splitters>=0.2.0,<0.4.0
faiss-cpu>=1.7.0
chardet>=5.0.0
tiktoken>=0.5.0

# Optional packages (install separately if needed):
# xhtml2pdf>=0.2.11  # For PDF output generation (requires: brew install cairo pkg-config)
# pymupdf>=1.23.0    # Alternative PDF parser (requires: brew install cairo pkg-config)
